chore: move `paragraph` stories to `paragraph-docs` package
