+++ .changeset/hip-views-write.md
@@ -1,0 +1,5 @@
+---
+'@nl-design-system-candidate/paragraph-docs': minor
+---
+
+Stories toegevoegd als exports zodat deze kunnen worden hergebruikt
