+++ packages/docs/paragraph-docs/package.json
@@ -11,7 +11,8 @@
     "paragraph-component"
   ],
   "files": [
-    "./docs/"
+    "./docs/",
+    "./stories/"
   ],
   "repository": {
     "type": "git+ssh",
@@ -21,5 +22,9 @@
   "publishConfig": {
     "access": "public",
     "provenance": true
+  },
+  "devDependencies": {
+    "@nl-design-system-candidate/paragraph-react": "workspace:*",
+    "@storybook/react-vite": "9.1.5"
   }
 }
