+++ packages/storybook/package.json
@@ -30,6 +30,7 @@
     "@nl-design-system-candidate/mark-docs": "workspace:*",
     "@nl-design-system-candidate/link-docs": "workspace:*",
     "@nl-design-system-candidate/link-react": "workspace:*",
+    "@nl-design-system-candidate/paragraph-docs": "workspace:*",
     "@nl-design-system-candidate/paragraph-react": "workspace:*",
     "@nl-design-system-candidate/skip-link-docs": "workspace:*",
     "@nl-design-system-candidate/storybook-shared": "workspace:*",
