+++ packages/storybook/stories/paragraph.stories.tsx
@@ -1,20 +1,17 @@
 import { ExampleBodyTextDecorator } from '@nl-design-system-candidate/storybook-shared/src/ExampleBodyTextDecorator';
-import type { Meta, StoryObj } from '@storybook/react-vite';
-import '../../components-css/paragraph-css/src/paragraph.scss';
+import type { Meta } from '@storybook/react-vite';
 import packageJSON from '../../components-react/paragraph-react/package.json';
-import { Paragraph } from '../../components-react/paragraph-react/src/paragraph';
+import { Paragraph as ParagraphComponent } from '@nl-design-system-candidate/paragraph-react';
+import paragraphMeta from '@nl-design-system-candidate/paragraph-docs/stories/paragraph.react.meta';
+import * as Stories from '@nl-design-system-candidate/paragraph-docs/stories/paragraph.stories';
 
 const meta = {
-  argTypes: {
-    children: { table: { category: 'API' } },
-    purpose: {
-      control: { labels: { undefined: '(undefined)' }, type: 'select' },
-      options: [undefined, 'lead'],
-      table: { category: 'API' },
-    },
-  },
-  component: Paragraph,
+  ...paragraphMeta,
   decorators: [ExampleBodyTextDecorator],
+  globals: {
+    dir: 'ltr',
+    lang: 'nl',
+  },
   parameters: {
     externalLinks: [
       {
@@ -28,32 +25,9 @@ const meta = {
     ],
   },
   title: 'Paragraph',
-} satisfies Meta<typeof Paragraph>;
+} satisfies Meta<typeof ParagraphComponent>;
 
 export default meta;
 
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  name: 'Paragraph',
-  args: {
-    children: 'Op brute wĳze ving de schooljuf de quasi-kalme lynx.',
-    purpose: undefined,
-  },
-  globals: {
-    dir: 'ltr',
-    lang: 'nl',
-  },
-};
-
-export const Lead: Story = {
-  name: 'Paragraph Lead',
-  args: {
-    children: 'Op brute wĳze ving de schooljuf de quasi-kalme lynx.',
-    purpose: 'lead',
-  },
-  globals: {
-    dir: 'ltr',
-    lang: 'nl',
-  },
-};
+export const Paragraph = Stories.Paragraph;
+export const Lead = Stories.Lead;
